Added method for sorted print
This method prints in the same format
as PrettyPrinter but sorts the keys
of each object lexicographically.

The impetus for this change was this
question:
http://stackoverflow.com/questions/31418626/sort-fields-in-rendered-json

In general it is useful to be able to have more deterministic
control over the ordering of output, if diff tools
are to be used.