Deprecate old parsing infrastructure and provide new entry points
The main entry points for parsing are now

 * spray.json.parseJsonAs -> parse JSON document to Scala object T directly
 * spray.json.parseJson -> parse JSON document to AST

There are also enrichments for the supported parser input types.

Inputs themselves are now closed down to String, Array[Char], Array[Byte], and
a more generic RandomAccessBytes that can be implemented by third parties.

Internally, we still use the current JsonParser infrastructure but that
will go (or not be used by default any more) in the future.